move rdf-terms utilities into a dedicated namespace
