+++ examples/custom-functions.js
@@ -1,8 +1,7 @@
 'use strict'
 
 const { Parser, Store } = require('n3')
-const { HashMapDataset, Graph, PlanBuilder } = require('../')
-const terms = require('../dist/rdf-terms')
+const { HashMapDataset, Graph, PlanBuilder, terms } = require('../')
 
 // Format a triple pattern according to N3 API:
 // SPARQL variables must be replaced by `null` values
